Update README.md
Added Map mode in the right corner to allow switching from vanilla to namalsk formats. 
Added search bar filter to allow searching different field instead of only name
Added the ability to clock the field names at the top to change the sort and a button to reset said sort.